# marketAPI/docs/swagger.yaml
basePath: /
definitions:
  domain.Ad:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      is_owner:
        type: boolean
      price:
        type: number
      title:
        type: string
      username:
        description: 'Для ответа:'
        type: string
    type: object
  domain.AdRequest:
    properties:
      description:
        maxLength: 1000
        minLength: 10
        type: string
      image_url:
        type: string
      price:
        minimum: 0
        type: number
      title:
        maxLength: 100
        minLength: 5
        type: string
    required:
    - description
    - image_url
    - price
    - title
    type: object
  domain.AuthRequest:
    properties:
      password:
        maxLength: 72
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  domain.AuthResponse:
    properties:
      token:
        type: string
    type: object
  domain.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  domain.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for marketplace application
  title: Marketplace API
  version: "1.0"
paths:
  /ads:
    get:
      consumes:
      - application/json
      description: Get list of advertisements with filtering and pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - default: date
        description: Sort field (date or price)
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order (asc or desc)
        in: query
        name: sort_order
        type: string
      - description: Minimum price
        in: query
        name: min_price
        type: number
      - description: Maximum price
        in: query
        name: max_price
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Ad'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: List ads
      tags:
      - ads
    post:
      consumes:
      - application/json
      description: Create new advertisement
      parameters:
      - description: Ad data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.AdRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Ad'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create new ad
      tags:
      - ads
  /ads/{id}:
    get:
      consumes:
      - application/json
      description: Get advertisement details by ID
      parameters:
      - description: Ad ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Ad'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get ad by ID
      tags:
      - ads
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login user with username and password
      parameters:
      - description: User credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register new user with username and password
      parameters:
      - description: User credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.AuthRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Register new user
      tags:
      - auth
  /health:
    get:
      consumes:
      - application/json
      description: Возвращает статус работы API и подключения к БД
      produces:
      - application/json
      responses:
        "200":
          description: Сервис работает
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Проверка работоспособности сервиса
      tags:
      - service
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"